// src/models/Job.ts
export type Job = {
  id: string
  timestamp: string
  companyname: string
  address: string
  websiteurl: string
  dentonchamberofcommercemember: string
  emailaddress: string
  phonenumber: string
  jobtitle: string
  jobdescription: string
  typeofwork: string
  jobcategory: string
  jobrate: string
  jobpostlink: string
  approved: string
}